Refactor Device to improve readability
With a split of packet processing logic across platforms (Apple and
non-Apple) the code quickly became bloated with cfg gating. The purpose
of this commit is to reduce the cfg attribute usage and make the
distinction between different packet processing approaches more clear.

The idea is to introduce packet_workers module that will encapsulate the
elements related to packet processing on worker threads, as used by
non-Apple platforms
